fix: fixing csv transform
