--- datasets/fec/pipelines/_images/run_csv_transform_kub_2/requirements.txt
@@ -1,3 +1,0 @@
-google-cloud-storage
-pandas
-requests
