+++ datasets/fec/pipelines/individuals_2016/individuals_2016_dag.py
@@ -37,7 +37,7 @@ with DAG(
     # Task to copy `individuals` to gcs
     download_zip_file = bash.BashOperator(
         task_id="download_zip_file",
-        bash_command='mkdir -p $data_dir/individuals\ncurl -o $data_dir/individuals/indiv16.zip -L $fec\nunzip $data_dir/individuals/indiv16.zip -d $data_dir/individuals/\nawk \u0027NR%7000000==1{x="/home/airflow/gcs/data/fec/individuals/F"++i;}{print \u003e x}\u0027  $data_dir/individuals/itcont.txt\n',
+        bash_command='mkdir -p $data_dir/individuals\ncurl -o $data_dir/individuals/indiv16.zip -L $fec\nunzip $data_dir/individuals/indiv16.zip -d $data_dir/individuals/\nsplit -l 7000000 --additional-suffix=.txt "$data_dir/individuals/itcont.txt" "$data_dir/individuals/F_"\n',
         env={
             "data_dir": "/home/airflow/gcs/data/fec",
             "fec": "https://cg-519a459a-0ea3-42c2-b7bc-fa1143481f74.s3-us-gov-west-1.amazonaws.com/bulk-downloads/2016/indiv16.zip",
@@ -52,15 +52,15 @@ with DAG(
         namespace="composer",
         service_account_name="datasets",
         image_pull_policy="Always",
-        image="{{ var.json.fec.container_registry.run_csv_transform_kub_2 }}",
+        image="{{ var.json.fec.container_registry.run_csv_transform_kub }}",
         env_vars={
             "SOURCE_GCS_BUCKET": "{{ var.value.composer_bucket }}",
-            "SOURCE_GCS_OBJECT": "data/fec/individuals/F1.txt",
+            "SOURCE_GCS_OBJECT": "data/fec/individuals/F_aa.txt",
             "SOURCE_FILE": "files/F1.txt",
             "TARGET_FILE": "files/data_output.csv",
             "TARGET_GCS_BUCKET": "{{ var.value.composer_bucket }}",
             "TARGET_GCS_PATH": "data/fec/individuals/data_output_1.csv",
-            "TABLE_NAME": "individuals_2016",
+            "PIPELINE_NAME": "individuals_2016",
             "CSV_HEADERS": '["cmte_id","amndt_ind","rpt_tp","transaction_pgi","image_num","transaction_tp","entity_tp","name","city","state", "zip_code","employer","occupation","transaction_dt","transaction_amt","other_id","tran_id","file_num", "memo_cd","memo_text","sub_id"]',
             "RENAME_MAPPINGS": '{"0":"cmte_id","1":"amndt_ind","2":"rpt_tp","3":"transaction_pgi","4":"image_num","5":"transaction_tp", "6":"entity_tp","7":"name","8":"city","9":"state","10":"zip_code","11":"employer", "12":"occupation","13":"transaction_dt","14":"transaction_amt","15":"other_id","16":"tran_id", "17":"file_num","18":"memo_cd","19":"memo_text","20":"sub_id"}',
         },
@@ -219,15 +219,15 @@ with DAG(
         namespace="composer",
         service_account_name="datasets",
         image_pull_policy="Always",
-        image="{{ var.json.fec.container_registry.run_csv_transform_kub_2 }}",
+        image="{{ var.json.fec.container_registry.run_csv_transform_kub }}",
         env_vars={
             "SOURCE_GCS_BUCKET": "{{ var.value.composer_bucket }}",
-            "SOURCE_GCS_OBJECT": "data/fec/individuals/F2.txt",
+            "SOURCE_GCS_OBJECT": "data/fec/individuals/F_ab.txt",
             "SOURCE_FILE": "files/F2.txt",
             "TARGET_FILE": "files/data_output.csv",
             "TARGET_GCS_BUCKET": "{{ var.value.composer_bucket }}",
             "TARGET_GCS_PATH": "data/fec/individuals/data_output_2.csv",
-            "TABLE_NAME": "individuals_2016",
+            "PIPELINE_NAMEME": "individuals_2016",
             "CSV_HEADERS": '["cmte_id","amndt_ind","rpt_tp","transaction_pgi","image_num","transaction_tp","entity_tp","name","city","state", "zip_code","employer","occupation","transaction_dt","transaction_amt","other_id","tran_id","file_num", "memo_cd","memo_text","sub_id"]',
             "RENAME_MAPPINGS": '{"0":"cmte_id","1":"amndt_ind","2":"rpt_tp","3":"transaction_pgi","4":"image_num","5":"transaction_tp", "6":"entity_tp","7":"name","8":"city","9":"state","10":"zip_code","11":"employer", "12":"occupation","13":"transaction_dt","14":"transaction_amt","15":"other_id","16":"tran_id", "17":"file_num","18":"memo_cd","19":"memo_text","20":"sub_id"}',
         },
@@ -381,15 +381,15 @@ with DAG(
         namespace="composer",
         service_account_name="datasets",
         image_pull_policy="Always",
-        image="{{ var.json.fec.container_registry.run_csv_transform_kub_2 }}",
+        image="{{ var.json.fec.container_registry.run_csv_transform_kub }}",
         env_vars={
             "SOURCE_GCS_BUCKET": "{{ var.value.composer_bucket }}",
-            "SOURCE_GCS_OBJECT": "data/fec/individuals/F3.txt",
+            "SOURCE_GCS_OBJECT": "data/fec/individuals/F_ac.txt",
             "SOURCE_FILE": "files/F3.txt",
             "TARGET_FILE": "files/data_output.csv",
             "TARGET_GCS_BUCKET": "{{ var.value.composer_bucket }}",
             "TARGET_GCS_PATH": "data/fec/individuals/data_output_3.csv",
-            "TABLE_NAME": "individuals_2016",
+            "PIPELINE_NAME": "individuals_2016",
             "CSV_HEADERS": '["cmte_id","amndt_ind","rpt_tp","transaction_pgi","image_num","transaction_tp","entity_tp","name","city","state", "zip_code","employer","occupation","transaction_dt","transaction_amt","other_id","tran_id","file_num", "memo_cd","memo_text","sub_id"]',
             "RENAME_MAPPINGS": '{"0":"cmte_id","1":"amndt_ind","2":"rpt_tp","3":"transaction_pgi","4":"image_num","5":"transaction_tp", "6":"entity_tp","7":"name","8":"city","9":"state","10":"zip_code","11":"employer", "12":"occupation","13":"transaction_dt","14":"transaction_amt","15":"other_id","16":"tran_id", "17":"file_num","18":"memo_cd","19":"memo_text","20":"sub_id"}',
         },
