+++ datasets/fec/pipelines/individuals_ingest_2020/pipeline.yaml
@@ -67,7 +67,7 @@ dag:
           TARGET_FILE: "files/data_output.csv"
           TARGET_GCS_BUCKET: "{{ var.value.composer_bucket }}"
           TARGET_GCS_PATH: "data/fec/individuals/data_output_ingest.csv"
-          TABLE_NAME: "individuals_ingest_2020"
+          PIPELINE_NAME: "individuals_ingest_2020"
           CSV_HEADERS: >-
             ["cmte_id","amndt_ind","rpt_tp","transaction_pgi","image_num","transaction_tp","entity_tp","name","city","state",
             "zip_code","employer","occupation""transaction_dt","transaction_amt","other_id","tran_id","file_num",
